using byte hours for at rest tally

+++ satellite/satellitedb/locked.go
@@ -124,7 +124,7 @@ func (m *lockedAccounting) LastRawTime(ctx context.Context, timestampType string
 }
 
 // SaveAtRestRaw records raw tallies of at-rest-data.
-func (m *lockedAccounting) SaveAtRestRaw(ctx context.Context, latestTally time.Time, nodeData map[storj.NodeID]int64) error {
+func (m *lockedAccounting) SaveAtRestRaw(ctx context.Context, latestTally time.Time, nodeData map[storj.NodeID]float64) error {
 	m.Lock()
 	defer m.Unlock()
 	return m.db.SaveAtRestRaw(ctx, latestTally, nodeData)
